use proper type casts for isspace() arguments
git-svn-id: https://svn.code.sf.net/p/gphoto/code/trunk/gphoto2@7894 67ed7778-7388-44ab-90cf-0a291f65f57c

+++ gphoto2/shell.c
@@ -133,11 +133,11 @@ shell_arg_count (const char *args)
 	int x=0, in_arg=0, count=0;
 	
 	while (x < strlen (args)) {
-		if ((!isspace(args[x])) && (!in_arg)) {
+		if ((!isspace((int)(args[x]))) && (!in_arg)) {
 			in_arg = 1;
 			count++;
 		}
-		if ((isspace(args[x])) && (in_arg))
+		if ((isspace((int)(args[x]))) && (in_arg))
 			in_arg = 0;
 		x++;
 	}
@@ -195,13 +195,13 @@ shell_arg (const char *args, int arg_num, char *arg)
 		return (GP_ERROR);
 
 	while (x < strlen(args)) {				/* Edge-triggered */
-		if ((!isspace(args[x])) && (!in_arg)) {
+		if ((!isspace((int)(args[x]))) && (!in_arg)) {
 			in_arg = 1;
 			if (count == arg_num)
 				copy = 1;
 			count++;
 		}
-		if ((isspace(args[x])) && (in_arg)) {
+		if ((isspace((int)(args[x]))) && (in_arg)) {
 			copy = 0;
 			in_arg = 0;
 		}
